Allow specifying custom FilePath for recording
If FilePath is not specified, the original behavior will be the
default.

This change does alter the overridding behavior for properties such
as PreferredSampleRate on Android as it previously would allow
the user's value to override what the Android platform was. The new
behavior would result in the Android platform value being evaluated
after the user's value had been set (rather than before).